+++ laskari2/Verkkokauppa1/src/main/java/ohtu/Main.java
@@ -1,13 +1,18 @@
 package ohtu;
 
-import ohtu.verkkokauppa.Kauppa;
-import ohtu.verkkokauppa.Kirjanpito;
+import ohtu.verkkokauppa.*;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.support.FileSystemXmlApplicationContext;
 
 public class Main {
 
     public static void main(String[] args) {
-        Kauppa kauppa = new Kauppa();
-
+        ApplicationContext ctx = new FileSystemXmlApplicationContext("src/main/resources/spring-context.xml");
+        //Kirjanpito kirjanpito = new Kirjanpito_laMaestro();
+        //Kauppa kauppa = new Kauppa(new Varasto_Espoo(kirjanpito), new Pankki_Nordea(kirjanpito), new ViitegeneraattoriDeluxe() );
+        //Kauppa kauppa = new Kauppa();
+        Kauppa kauppa = ctx.getBean(Kauppa.class);
+        Kirjanpito kirjanpito = ctx.getBean(Kirjanpito_laMaestro.class);
         // kauppa hoitaa yhden asiakkaan kerrallaan seuraavaan tapaan:
         kauppa.aloitaAsiointi();
         kauppa.lisaaKoriin(1);
@@ -25,7 +30,7 @@ public class Main {
         kauppa.tilimaksu("Arto Vihavainen", "3425-1652");
 
         // kirjanpito
-        for (String tapahtuma : Kirjanpito.getInstance().getTapahtumat()) {
+        for (String tapahtuma : kirjanpito.getTapahtumat()) {
             System.out.println(tapahtuma);
         }
     }
