+++ laskari2/Verkkokauppa1/src/main/java/ohtu/verkkokauppa/Kirjanpito.java
@@ -3,28 +3,9 @@ package ohtu.verkkokauppa;
 
 import java.util.ArrayList;
 
-public class Kirjanpito {
-    private static Kirjanpito instance;
-    
-    public static Kirjanpito getInstance() {
-        if ( instance==null) {
-            instance = new Kirjanpito();
-        }
-        
-        return instance;
-    }
-    
-    private ArrayList<String> tapahtumat;
+public interface Kirjanpito {
 
-    public Kirjanpito() {
-        tapahtumat = new ArrayList<String>();
-    }
-    
-    public void lisaaTapahtuma(String tapahtuma) {
-        tapahtumat.add(tapahtuma);
-    }
+    public void lisaaTapahtuma(String tapahtuma);
 
-    public ArrayList<String> getTapahtumat() {
-        return tapahtumat;
-    }       
+    public ArrayList<String> getTapahtumat();
 }
