+++ laskari2/Verkkokauppa1/src/main/java/ohtu/verkkokauppa/Ostoskori.java
@@ -1,7 +1,10 @@
 package ohtu.verkkokauppa;
 
+import org.springframework.stereotype.Component;
+
 import java.util.ArrayList;
 
+@Component
 public class Ostoskori {
 
     private ArrayList<Tuote> tuotteet;
