+++ laskari2/Verkkokauppa1/src/main/java/ohtu/verkkokauppa/Pankki.java
@@ -1,27 +1,6 @@
 package ohtu.verkkokauppa;
 
-public class Pankki {
+public interface Pankki {
 
-    private static Pankki instanssi;
-
-    public static Pankki getInstance() {
-        if (instanssi == null) {
-            instanssi = new Pankki();
-        }
-
-        return instanssi;
-    }
-    private Kirjanpito kirjanpito;
-
-    public Pankki() {
-        kirjanpito = Kirjanpito.getInstance();
-    }
-
-    public boolean tilisiirto(String nimi, int viitenumero, String tililta, String tilille, int summa) {
-        kirjanpito.lisaaTapahtuma("tilisiirto: tililtä " + tilille + " tilille " + tilille
-                + " viite " + viitenumero + " summa " + summa + "e");
-        
-        // täällä olisi koodi joka ottaa yhteyden pankin verkkorajapintaan
-        return true;
-    }
+    public boolean tilisiirto(String nimi, int viitenumero, String tililta, String tilille, int summa);
 }
