+++ laskari2/Verkkokauppa1/src/main/java/ohtu/verkkokauppa/Tuote.java
@@ -1,6 +1,9 @@
 
 package ohtu.verkkokauppa;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+
 public class Tuote {
     private int id;
     private String nimi;
