+++ laskari2/Verkkokauppa1/src/main/java/ohtu/verkkokauppa/Viitegeneraattori.java
@@ -1,24 +1,6 @@
 package ohtu.verkkokauppa;
 
-public class Viitegeneraattori {
-
-    private static Viitegeneraattori instanssi;
-
-    public static Viitegeneraattori getInstance() {
-        if (instanssi == null) {
-            instanssi = new Viitegeneraattori();
-        }
-
-        return instanssi;
-    }
+public interface Viitegeneraattori {
     
-    private int seuraava;
-    
-    private Viitegeneraattori(){
-        seuraava = 1;    
-    }
-    
-    public int uusi(){
-        return seuraava++;
-    }
+    public int uusi();
 }
